# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Shop API",
        "version": "1.0"
    },
    "paths": {
        "/api/add_admin": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Добавить админа",
                "operationId": "99e5eaf22fded43af6bee0135f370254",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin added successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/delete_admin": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Удалить админа",
                "operationId": "8e3cda48c7481f442f56d177964d1db2",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Андрей"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Лясковский"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "1-й Красногвардейский проезд, 22с1, Москва"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+7 (903) 111-11-11"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "lae3145@mail.ru"
                                            },
                                            "email_verified_at": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Информаиця о пользователе",
                "operationId": "f0b15943f55143d36f606a2dbd1d2366",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Андрей"
                                            },
                                            "surname": {
                                                "type": "string",
                                                "example": "Лясковский"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "1-й Красногвардейский проезд, 22с1, Москва"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "+7 (903) 111-11-11"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "lae3145@mail.ru"
                                            },
                                            "email_verified_at": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "created_at": {
                                                "type": "time",
                                                "example": "2023-07-06T08:27:30.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "time",
                                                "example": "2023-07-06T09:45:07.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Обновление токена",
                "operationId": "b2eae7a8032954d25b46c7ec6f489c68",
                "responses": {
                    "200": {
                        "description": "Successfully refresh"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Your token",
                "name": "Auth",
                "in": "header"
            }
        }
    }
}